Refactor Ratmas commands into separate files for better organization
